backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==2.1.1
openai==1.3.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0